Change background Color and game difficulty
I added a feature in which, when a certain number of points are scored, the color of the background changes. Also added a level of difficulty. When we start the game, the speed of the obstacles slowly begins to accelerate, and when we score a certain number of points, the obstacles will change and more difficult obstacles will be encountered